# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Middleware API",
...
    "license": {
      "name": "MIT"
    },
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "https://api.thisisatest.tech",
      "description": "Public API (set PUBLIC_URL)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/mcp/tools.json": {
      "get": {
        "tags": [
          "MCP"
        ],
        "summary": "Mcp Tools Manifest",
        "description": "Return the MCP tools.json manifest.\n\nThis is an unauthenticated HTTP mirror of tool metadata exposed by this\nproject's governed JSON-RPC subset. Clients can obtain the same list by\nsending ``tools/list`` to ``/mcp/messages``; that endpoint does not\nimplement the complete MCP initialization lifecycle.\n\nQuery Parameters:\n    category: Optional service category filter\n\nReturns:\n    MCP tools.json manifest with tool definitions",
        "operationId": "MCP_Tools_Manifest_mcp_tools_json_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ServiceCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/messages": {
      "post": {
        "tags": [
          "MCP"
        ],
        "summary": "Mcp Json-Rpc Messages",
        "description": "Legacy project JSON-RPC endpoint, kept for existing clients and the local proof scripts. It does not implement the standard MCP initialization lifecycle. Check `/.well-known/agent.json` before assuming the standard MCP Streamable HTTP endpoint at POST /mcp is available; both transports run the same governed permit→meter→receipt path when enabled.",
        "operationId": "MCP_JSON_RPC_Messages_mcp_messages_post",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/discover/tools": {
      "get": {
        "tags": [
          "Agent Discovery"
        ],
        "summary": "List Available MCP Tools",
        "description": "Returns all available MCP tools with their schemas and pricing.",
        "operationId": "list_mcp_tools_v1_discover_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceCategory": {
        "type": "string",
        "enum": [
          "iot_bridge",
          "telemetry_pm",
          "media_engine",
          "agent_comms",
          "content_factory",
          "red_team",
          "oracle",
          "platform_fee",
          "swarm_delegation",
          "protocol_gen",
          "sandbox",
          "rtaas"
        ],
        "title": "ServiceCategory"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for agent authentication. Pass in the X-API-Key header.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}